criando as tabelas e os types
