+++ semana20/labook-projeto/template-labook/src/data/Migrations.ts
@@ -1,0 +1,30 @@
+import { BaseDataBase } from "./BaseDataBase";
+
+export class Migration extends BaseDataBase{
+
+   public migration = async () => {
+      await BaseDataBase.connection.raw(`
+         CREATE TABLE IF NOT EXISTS labook_users(
+            id VARCHAR(255) PRIMARY KEY,
+            name VARCHAR(255) NOT NULL,
+            email VARCHAR(255) UNIQUE NOT NULL,
+            password VARCHAR(255) NOT NULL
+         );
+
+         CREATE TABLE IF NOT EXISTS labook_posts(
+            id VARCHAR(255) PRIMARY KEY,
+            photo VARCHAR(255) NOT NULL,
+            description VARCHAR(255) NOT NULL,
+            type ENUM("normal","event") DEFAULT "normal",
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            author_id VARCHAR(255),
+            FOREIGN KEY (author_id) REFERENCES labook_users (id)
+         )
+      `)
+      .then(console.log)
+      .catch(console.log)
+}}
+
+const mb = new Migration();
+
+const connect = mb.migration();
