+++ semana20/labook-projeto/template-labook/src/typePosts/typePosts.ts
@@ -1,0 +1,13 @@
+export enum TYPE_POST {
+    normal = "normal",
+    event = "event"
+}
+
+export type post = {
+   id: string,
+   photo: string,
+   description: string,
+   type: TYPE_POST,
+   createdAt: Date,
+   authorId: string
+}
