+++ semana20/labook-projeto/template-labook/src/typeUsers/types.ts
@@ -1,0 +1,12 @@
+export type user = {
+
+    id: string,
+    name: string,
+    email: string,
+    password: string
+
+}
+
+export type authenticationData = {
+    id: string
+}
