--- semana20/labook-projeto/template-labook/src/types.ts
@@ -1,1 +1,0 @@
-https://github.com/future4code/Micheline_Barros/pull/62
